+++ functions.php
@@ -10,27 +10,27 @@
  * @link    http://www.studiopress.com/
  */
 
-//* Start the engine
+// Start the engine.
 include_once( get_template_directory() . '/lib/init.php' );
 
-//* Setup Theme
+// Setup Theme.
 include_once( get_stylesheet_directory() . '/lib/theme-defaults.php' );
 
-//* Set Localization (do not remove)
+// Set Localization (do not remove).
 load_child_theme_textdomain( 'genesis-sample', apply_filters( 'child_theme_textdomain', get_stylesheet_directory() . '/languages', 'genesis-sample' ) );
 
-//* Add Image upload and Color select to WordPress Theme Customizer
+// Add Image upload and Color select to WordPress Theme Customizer.
 require_once( get_stylesheet_directory() . '/lib/customize.php' );
 
-//* Include Customizer CSS
+// Include Customizer CSS.
 include_once( get_stylesheet_directory() . '/lib/output.php' );
 
-//* Child theme (do not remove)
+// Child theme (do not remove).
 define( 'CHILD_THEME_NAME', 'Genesis Sample' );
 define( 'CHILD_THEME_URL', 'http://www.studiopress.com/' );
 define( 'CHILD_THEME_VERSION', '2.3.0' );
 
-//* Enqueue Scripts and Styles
+// Enqueue Scripts and Styles.
 add_action( 'wp_enqueue_scripts', 'genesis_sample_enqueue_scripts_styles' );
 function genesis_sample_enqueue_scripts_styles() {
 
@@ -46,7 +46,7 @@ function genesis_sample_enqueue_scripts_styles() {
 
 }
 
-//* Enqueue custom WooCommerce styles when WooCommerce active
+// Enqueue custom WooCommerce styles when WooCommerce active.
 add_filter( 'woocommerce_enqueue_styles', 'genesis_sample_woocommerce_styles' );
 function genesis_sample_woocommerce_styles( $enqueue_styles ) {
 
@@ -61,19 +61,19 @@ function genesis_sample_woocommerce_styles( $enqueue_styles ) {
 
 }
 
-//* Add WooCommerce Support
+// Add WooCommerce Support.
 add_theme_support( 'genesis-connect-woocommerce' );
 
-//* Add HTML5 markup structure
+// Add HTML5 markup structure.
 add_theme_support( 'html5', array( 'caption', 'comment-form', 'comment-list', 'gallery', 'search-form' ) );
 
-//* Add Accessibility support
+// Add Accessibility support.
 add_theme_support( 'genesis-accessibility', array( '404-page', 'drop-down-menu', 'headings', 'rems', 'search-form', 'skip-links' ) );
 
-//* Add viewport meta tag for mobile browsers
+// Add viewport meta tag for mobile browsers.
 add_theme_support( 'genesis-responsive-viewport' );
 
-//* Add support for custom header
+// Add support for custom header.
 add_theme_support( 'custom-header', array(
 	'width'           => 600,
 	'height'          => 160,
@@ -82,26 +82,26 @@ add_theme_support( 'custom-header', array(
 	'flex-height'     => true,
 ) );
 
-//* Add support for custom background
+// Add support for custom background.
 add_theme_support( 'custom-background' );
 
-//* Add support for after entry widget
+// Add support for after entry widget.
 add_theme_support( 'genesis-after-entry-widget-area' );
 
-//* Add support for 3-column footer widgets
+// Add support for 3-column footer widgets.
 add_theme_support( 'genesis-footer-widgets', 3 );
 
-//* Add Image Sizes
+// Add Image Sizes.
 add_image_size( 'featured-image', 720, 400, TRUE );
 
-//* Rename primary and secondary navigation menus
+// Rename primary and secondary navigation menus.
 add_theme_support( 'genesis-menus' , array( 'primary' => __( 'After Header Menu', 'genesis-sample' ), 'secondary' => __( 'Footer Menu', 'genesis-sample' ) ) );
 
-//* Reposition the secondary navigation menu
+// Reposition the secondary navigation menu.
 remove_action( 'genesis_after_header', 'genesis_do_subnav' );
 add_action( 'genesis_footer', 'genesis_do_subnav', 5 );
 
-//* Reduce the secondary navigation menu to one level depth
+// Reduce the secondary navigation menu to one level depth.
 add_filter( 'wp_nav_menu_args', 'genesis_sample_secondary_menu_args' );
 function genesis_sample_secondary_menu_args( $args ) {
 
@@ -115,7 +115,7 @@ function genesis_sample_secondary_menu_args( $args ) {
 
 }
 
-//* Modify size of the Gravatar in the author box
+// Modify size of the Gravatar in the author box.
 add_filter( 'genesis_author_box_gravatar_size', 'genesis_sample_author_box_gravatar' );
 function genesis_sample_author_box_gravatar( $size ) {
 
@@ -123,7 +123,7 @@ function genesis_sample_author_box_gravatar( $size ) {
 
 }
 
-//* Modify size of the Gravatar in the entry comments
+// Modify size of the Gravatar in the entry comments.
 add_filter( 'genesis_comment_list_args', 'genesis_sample_comments_gravatar' );
 function genesis_sample_comments_gravatar( $args ) {