+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][unreleased]
 * Add WooCommerce support to the theme and customizer
+* Update Documentation
 
 ## [2.2.4] - 2016-06-08
 * Reorder font size on breadcrumbs