+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * Add WooCommerce support to the theme and customizer
 * Add an updated and improved responsive menu script
 * Updated the file structure
-* Update Documentation
+* Updated documentation
+* Updated code standards
 
 ## [2.2.4] - 2016-06-08
 * Reorder font size on breadcrumbs