+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Genesis Sample Theme Changelog
 
-## [Unreleased][unreleased]
+## [2.3.0] - 2017-01-30
 * Add WooCommerce support to the theme and customizer
 * Add an updated and improved responsive menu script
 * Updated the file structure
@@ -27,6 +27,6 @@
 * Set localization
 * Update XML file
 
-[unreleased]: https://github.com/copyblogger/genesis-sample/compare/2.2.4...HEAD
+[2.3.0]: https://github.com/copyblogger/genesis-sample/compare/2.2.4...2.3.0
 [2.2.4]: https://github.com/copyblogger/genesis-sample/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/copyblogger/genesis-sample/compare/014deb3689323b7bbd4ddbfff4f5f9279a38f741...2.2.3